test: run Nightly asset contracts in reduced CI
