+++ tests/test_nightly_asset_contracts.py
@@ -1,0 +1,78 @@
+"""Coverage contracts for committed assets consumed by Nightly scans."""
+
+from pathlib import Path
+
+import pytest
+
+from modelaudit.core import determine_exit_code, scan_model_directory_or_file
+from modelaudit.core_results import results_have_inconclusive_outcome
+from modelaudit.models import ModelAuditResultModel
+
+ASSETS = Path(__file__).parent / "assets" / "samples" / "pickles"
+
+
+def _scan_asset(name: str) -> ModelAuditResultModel:
+    return scan_model_directory_or_file(str(ASSETS / name), cache_enabled=False)
+
+
+def test_complete_benign_asset_scans_cleanly() -> None:
+    """The committed benign fixture must remain complete and issue-free."""
+    result = _scan_asset("safe_data.pkl")
+
+    assert result.has_errors is False
+    assert results_have_inconclusive_outcome(result) is False
+    assert determine_exit_code(result) == 0
+    assert result.issues == []
+
+
+@pytest.mark.parametrize(
+    ("name", "rule_code", "message"),
+    [
+        (
+            "evil.pickle",
+            "S201",
+            "Found REDUCE opcode invoking dangerous global: posix.system",
+        ),
+        (
+            "malicious_model_realistic.pkl",
+            "S601",
+            "Encoded pickle payload detected",
+        ),
+        (
+            "nested_pickle_base64.pkl",
+            "S601",
+            "Encoded pickle payload detected",
+        ),
+    ],
+)
+def test_complete_malicious_assets_preserve_security_signal(
+    name: str,
+    rule_code: str,
+    message: str,
+) -> None:
+    """Nightly benchmark inputs must remain complete and security-positive."""
+    result = _scan_asset(name)
+
+    assert result.has_errors is False
+    assert results_have_inconclusive_outcome(result) is False
+    assert determine_exit_code(result) == 1
+    assert any(issue.rule_code == rule_code and issue.message == message for issue in result.issues)
+
+
+def test_intentional_incomplete_pickle_asset_preserves_security_exit() -> None:
+    """The dill fixture is intentionally incomplete but still security-positive."""
+    result = _scan_asset("dill_func.pkl")
+    path = str(ASSETS / "dill_func.pkl")
+    metadata = result.file_metadata[path].model_dump(exclude_none=True)
+
+    assert result.has_errors is False
+    assert results_have_inconclusive_outcome(result) is True
+    assert metadata["scan_outcome"] == "inconclusive"
+    assert metadata["scan_outcome_reasons"] == ["nested_pickle_incomplete"]
+    assert any(issue.message == "Nested pickle analysis did not complete" for issue in result.issues)
+    assert any(
+        issue.rule_code == "S201"
+        and issue.message == "Found REDUCE opcode invoking dangerous global: dill._dill._create_code"
+        for issue in result.issues
+    )
+    assert determine_exit_code(result) == 1
